Fix aggregate datastream validation
Join the path prefix with the payload key when looking for the mappings.
Moreover, remove the keySet check since it's currently wrong and redundant (we
already check that the payload and the mappings are the same length and then we
go through each payload key and search it, prefixed, in the mappings).

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>