Add a check on certificate availability before connect.
During the first connection ever the SDK always throws an exception because it tries to establish a connection to the MQTT before a certificate is issued.

The end goal is to prevent log pollution and fix a predictable exception.
To do so a check to `isCertificateAvalaible()` is now called before connecting and eventually, if it returns `false`, a new certificate is requested.
On top of that, a check on x509 certificate validity is inquired so a new certificate is requested even if the file is actually present but it's expired.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>