Rewrite registerDevice for Android API 23
Scanning the project I found out that in the method `registerDevice` the `Optional` class is used to describe a nullable object. Unluckily the `Optional` class was added in API Level 24 ([see documentation] (https://developer.android.com/reference/java/util/Optional#ofNullable(T)))  and it can't be used because `minSdkVersion 23` is declared in `DeviceSDKAndroid/build.gradle`.
This commit replaces the `Optional` class with a couple of not equally elegant `== null` checks.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>