Add dynamic interface load
Add the capability to load and unload interfaces dynamically in device's
introspection calling the `addInterface` and `removeInterface` functions
of the `AstarteDevice` class. The `AstartePairableDevice` class extends
said functions to add the capability to send the altered introspection
to the server if the device is connected. The function `loadInterface`
in the `InterfaceProvider` can be implemented to help load interfaces by
name and instanciate them to `JSONObject` similarly to the
`loadAllInterfaces` does for a batch of interfaces as shown in the
`ExampleInterfaceProvider` example implementation.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>