Add Sonatype publications to CI
Add a github action to automatize the publication of the Sonatype nexus staging repository.
The action will trigger when a release or draft of a release is published, or a pre-release is changed to a release.

Note: a manual interaction is still required to release the package from the staging repository to the public one. This step is deliberately left to an admin user because is not possible to remove a packet published by mistake.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>
Signed-off-by: Antonio Gisondi <antonio.gisondi@secomind.com>