Handle the sessionPresent flag from the broker
It looked like the SDK sent the introspection and all properties
every time the device connected even if it was a temporary
disconnection. The right behavior should be that those data should be
sent only on the first connection or if the session was cleaned by the
broker because considered invalid.
That anomaly was generated by the MQTT Client library cleaned
the session by default for each connection.
This commit deactivates this default behavior.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>