Add support to ignore_ssl_error flag
The SDK currently does not support self signed certificates on MQTT
connections, but this functionality is heavily needed for some self
hosted instances of Astarte.
This commit adds an optional flag in the AstarteDevice constructor that
enables a custom TrustManager that allows self signed certificates for
SSL sockets in paho and disables hostname verifications on mqtt
connection options.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>